+++ mimesis/schema.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
 """Implements classes for generating data by schema."""
-import warnings
 from typing import Any, Callable, Iterator, List, Optional
 
 from mimesis.exceptions import FieldError, SchemaError
@@ -159,40 +158,6 @@ class Schema:
         :return: List of fulfilled schemas.
         """
         return [self.schema() for _ in range(iterations)]
-
-    def loop(self) -> Iterator[JSON]:
-        """Fulfills a schema **infinitely** in a lazy way.
-
-        This method can be useful when you have some dynamic
-        conditions in depend on which the generation must be stopped.
-
-        Please, read all the notes and warnings below.
-
-        .. note::
-            Since data `mimesis` provides are limited, frequent calls of
-            this method can cause data duplication.
-
-        .. warning::
-            Do not use this method without interrupt conditions, otherwise,
-            you're risking running out of memory.
-
-        .. warning::
-            **Never** call `list()`, `tuple()` or any other callable which tries to
-            evaluate the whole lazy object on this method.
-
-            No, I'm serious, mate, **NEVER!**.
-
-        :return: An infinite iterator with fulfilled schemas.
-        """
-
-        warnings.warn(
-            "You're iterating over the infinite object! "
-            "The Schema.loop() can cause a serious memory leak."
-            "Please, see: https://mimesis.name/api.html#mimesis.schema.Schema.loop"
-        )
-
-        while True:
-            yield self.schema()
 
     def iterator(self, iterations: int = 1) -> Iterator[JSON]:
         """Fulfills schema in a lazy way.